+++ src/actinia_core/rest/download_cache_management.py
@@ -40,8 +40,8 @@ from actinia_core.core.common.exceptions import AsyncProcessError
 from actinia_core.models.response_models import \
     StorageResponseModel, StorageModel, ProcessingResponseModel
 from actinia_core.core.common.api_logger import log_api_call
-from actinia_core.core.user_auth import very_admin_role
-from actinia_core.core.user_auth import check_user_permissions
+from actinia_core.rest.base.user_auth import very_admin_role
+from actinia_core.rest.base.user_auth import check_user_permissions
 from actinia_core.core.common.app import auth
 
 __license__ = "GPLv3"
