+++ src/actinia_core/rest/ephemeral_custom_processing.py
@@ -31,7 +31,7 @@ from actinia_core.rest.ephemeral_processing import EphemeralProcessing
 from actinia_core.rest.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.exceptions import AsyncProcessError
-from actinia_core.core.user_auth import check_location_mapset_module_access
+from actinia_core.rest.base.user_auth import check_location_mapset_module_access
 from actinia_core.core.common.process_object import Process
 
 __license__ = "GPLv3"
