+++ src/actinia_core/rest/location_management.py
@@ -39,8 +39,8 @@ from actinia_api.swagger2.actinia_core.schemas.location_management \
 
 from actinia_core.core.common.app import auth
 from actinia_core.core.common.api_logger import log_api_call
-from actinia_core.core.user_auth import very_admin_role
-from actinia_core.core.user_auth import check_user_permissions
+from actinia_core.rest.base.user_auth import very_admin_role
+from actinia_core.rest.base.user_auth import check_user_permissions
 from actinia_core.models.response_models import ProcessingResponseModel
 from actinia_core.models.response_models import SimpleResponseModel
 from actinia_core.models.response_models import MapsetInfoResponseModel
