+++ src/actinia_core/rest/mapset_management.py
@@ -42,8 +42,8 @@ from actinia_core.core.common.app import auth
 from actinia_core.core.common.api_logger import log_api_call
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.exceptions import AsyncProcessError
-from actinia_core.core.user_auth import check_user_permissions
-from actinia_core.core.user_auth import very_admin_role
+from actinia_core.rest.base.user_auth import check_user_permissions
+from actinia_core.rest.base.user_auth import very_admin_role
 from actinia_core.models.response_models import ProcessingResponseModel, \
     StringListProcessingResultResponseModel, MapsetInfoResponseModel, \
     RegionModel, ProcessingErrorResponseModel
