+++ src/actinia_core/rest/mapsets.py
@@ -32,8 +32,8 @@ from flask import jsonify, make_response
 from flask_restful_swagger_2 import swagger
 from flask import request
 from .resource_base import ResourceBase
-from actinia_core.core.user_auth import check_user_permissions
-# from actinia_core.core.user_auth import very_admin_role
+from actinia_core.rest.base.user_auth import check_user_permissions
+# from actinia_core.rest.base.user_auth import very_admin_role
 from actinia_core.core.common.app import auth
 from actinia_core.core.common.api_logger import log_api_call
 from actinia_core.core.common.config import global_config
