+++ src/actinia_core/rest/resource_storage_management.py
@@ -43,8 +43,8 @@ from actinia_core.models.response_models import \
 
 from actinia_core.core.common.app import auth
 from actinia_core.core.common.api_logger import log_api_call
-from actinia_core.core.user_auth import very_admin_role
-from actinia_core.core.user_auth import check_user_permissions
+from actinia_core.rest.base.user_auth import very_admin_role
+from actinia_core.rest.base.user_auth import check_user_permissions
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert"
